refactor(gfn): compose WebRTC client from focused controllers
